+++ resources/views/components/ui/context-menu-sub.blade.php
@@ -2,8 +2,8 @@
     data-slot="context-menu-sub"
     x-data="blatMenu()"
     x-id="['blat-ctx-submenu', 'blat-ctx-submenu-trigger']"
-    @mouseenter="open = true"
-    @mouseleave="closeMenu(false)"
+    @mouseenter="open = true; cancelClose()"
+    @mouseleave="closeSoon()"
     {{ $attributes->twMerge('relative') }}
 >
     {{ $slot }}